% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_delta.R
\name{add_delta}
\alias{add_delta}
\title{Add delta variable}
\usage{
add_delta(
  data,
  cols,
  id,
  time,
  fill = 0,
  n = 1L,
  type = "lag",
  prefix = paste("delta", type, n, sep = "_")
)
}
\arguments{
\item{data}{the dataframe to be called}

\item{id}{the name of the id variable}

\item{var}{the variable name}
}
\description{
This function will create a delta variable given a subject id, variable name
and dataframe.
}
