% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Process_case_study.R
\name{Aggregate_Data}
\alias{Aggregate_Data}
\title{Aggregate Case Study Data}
\usage{
Aggregate_Data(village, Case_Study_Data, Case_Study_Sites)
}
\arguments{
\item{Case_Study_Data}{A data.frame}

\item{Sites}{Names of the village}
}
\value{
A list
}
\description{
Aggregate Case Study Data
}
