% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopDynamics.R
\name{Assess}
\alias{Assess}
\title{Title}
\usage{
Assess(
  data,
  options = list(),
  control = list(eval.max = 20000, iter.max = 20000, abs.tol = 1e-20),
  map = list(log_sigmaF = factor(NA), log_sigmaR = factor(NA), log_sigmaR0 =
    factor(NA)),
  Fit_Effort = 1,
  Fit_CPUE = 1,
  log_sigmaF = log(0.5),
  log_sigmaR = log(0.9),
  log_sigmaR0 = log(0.6)
)
}
\arguments{
\item{log_sigmaR0}{}
}
\description{
Title
}
