% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Report.R
\name{calc_average_monthly_F}
\alias{calc_average_monthly_F}
\alias{calc_average_monthly_SPR}
\alias{Calc_RP}
\alias{plot_RP}
\title{Calculate SPR and F reference points from HARA model}
\usage{
calc_average_monthly_F(assess)

calc_average_monthly_SPR(assess)

Calc_RP(assess, utilpow = c(0.4, 0.9))

plot_RP(assess, utilpow = c(0.4, 0.9))
}
\arguments{
\item{assess}{An object of class \code{Assess}}

\item{utilpow}{numeric vector of values for the HARA utility exponent}
}
\value{
A data.frame
}
\description{
Calculate SPR and F reference points from HARA model
}
\section{Functions}{
\itemize{
\item \code{calc_average_monthly_F()}: calculate average monthly F

\item \code{calc_average_monthly_SPR()}: calculate average monthly SPR

\item \code{plot_RP()}: plot the mean F and mean SPR and the optimal values from the HARA model

}}
