% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimize_F_Pattern.R
\name{calculate_optimal_fishing}
\alias{calculate_optimal_fishing}
\title{Calculate the optimal pattern in fishing mortality}
\usage{
calculate_optimal_fishing(
  LifeHistory,
  Exploitation,
  opt_type = 1,
  utilpow = 0.3
)
}
\arguments{
\item{LifeHistory}{A \code{LifeHistory} object}

\item{Exploitation}{An \code{Exploitation} object}

\item{opt_type}{0 to maximize yield, 1 to maximize HARA utility}

\item{utilpow}{Exponent of the HARA model}
}
\value{
A list
}
\description{
Uses a HARA model to calculate the optimal exploitation pattern
}
