% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaImport_From.R
\name{Check}
\alias{Check}
\title{Check a \code{Data} object}
\usage{
Check(Data = NULL, silent = FALSE)
}
\arguments{
\item{Data}{An object of class \code{Data}}
}
\value{
Nothing
}
\description{
Check a \code{Data} object
}
