% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Checks.R
\name{Check_Parameters}
\alias{Check_Parameters}
\title{Check Parameters}
\usage{
Check_Parameters(Parameters, Data, silent = FALSE)
}
\arguments{
\item{Parameters}{A \code{Parameters} object}

\item{Data}{A \code{Data} object}

\item{silent}{Logical. Hide the messages?}
}
\value{
Nothing. Returns an error if there is a problem with the Parameters
}
\description{
Check Parameters
}
