% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Assess.R
\name{Construct_Data_OM}
\alias{Construct_Data_OM}
\title{Title}
\usage{
Construct_Data_OM(
  SimMod,
  sim = 1,
  CAW_Monthly_ESS = 100,
  Effort_CV = 0.2,
  CPUE_CV = 0.2,
  Fit_Effort = 1,
  Fit_CPUE = 1,
  Fit_CAW = 1,
  use_Frwpen = 0,
  use_R0rwpen = 0,
  use_Fmeanprior = 0
)
}
\arguments{
\item{use_Fmeanprior}{}
}
\description{
Title
}
