% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_SLAM_gluc.R
\docType{data}
\name{data_SLAM_gluc}
\alias{data_SLAM_gluc}
\title{SLAM Glucose and Lactate Data}
\format{
a dataframe 17467 obs and 6 variables
\describe{
  \item{idno}{unique identifier}
  \item{date}{date of bleed}
  \item{gluc}{Glucose levels in mg/dL}
  \item{lact}{Lactate levels in mg/dL}
}
}
\usage{
data(data_SLAM_gluc)
}
\description{
A Dataset that has the glucose and lactate levels acquired from routine
bleeds in SLAM
}
\keyword{datasets}
