% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import_Data.R
\name{Example_Data}
\alias{Example_Data}
\title{Print the file path for example data sets included in SLAM}
\usage{
Example_Data()
}
\value{
The file paths for the example data
}
\description{
Print the file path for example data sets included in SLAM
}
\examples{
Example_Data()
}
