% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import_Parameters.R
\name{File_Location}
\alias{File_Location}
\title{Get the file path of the system files for the SLAM package}
\usage{
File_Location(file = "Simulation_Parameters.csv")
}
\arguments{
\item{file}{The name of the file to locate}
}
\value{
The location of the file included in the SLAM package
}
\description{
Get the file path of the system files for the SLAM package
}
\examples{
File_Location()
}
