% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulate.R
\name{Generate_Data}
\alias{Generate_Data}
\title{Generate Sampled Data from a Simulated Fishery}
\usage{
Generate_Data(Simulation = NULL, Sampling = NULL, seed = 101, silent = FALSE)
}
\arguments{
\item{Simulation}{An  object of class \code{Simulation} generated by \code{Simulate()}}

\item{Sampling}{A \code{list} object with the parameters for generating data}

\item{seed}{The seed for the random number generator}

\item{silent}{Hide messages?}
}
\description{
Generate Sampled Data from a Simulated Fishery
}
