% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generate_Monthly_Recruitment.R
\name{Generate_Monthly_Recruitment}
\alias{Generate_Monthly_Recruitment}
\title{Title}
\usage{
Generate_Monthly_Recruitment(mu = 4, sigma = 2)
}
\arguments{
\item{sigma}{}
}
\description{
Title
}
