% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import_Data.R
\name{Import}
\alias{Import}
\title{Import and organize data for assessment model}
\usage{
Import(xlfile, BinWidth = NULL, BinMax = NULL, silent = FALSE, ...)
}
\arguments{
\item{xlfile}{The full file path to an Excel file structured for SLAM data}

\item{BinWidth}{The width of the bins for the catch-at-weight data. Only needed if the CAW data is raw.}

\item{BinMax}{The maximum bin for the catch-at-weight data. Only needed if the CAW data is raw.}

\item{silent}{Logical. Hide the messages?}

\item{...}{Additional parameters (not currently used)}
}
\value{
A \code{Data} object for the \code{Assess} function
}
\description{
Import and organize data for assessment model
}
