% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import_Data.R
\name{Import_Data}
\alias{Import_Data}
\alias{Import_Data.Simulated}
\title{Import and organize data for assessment model}
\usage{
Import_Data(xlfile, BinWidth = NULL, use_Frwpen = 1, use_R0rwpen = 1, ...)

\method{Import_Data}{Simulated}(
  Sampled_Data = NULL,
  sim = 1,
  Fit_Effort = 1,
  Fit_CPUE = 1,
  Fit_CAW = 1,
  Fit_CAA = 1,
  use_Frwpen = 1,
  use_R0rwpen = 1,
  Data_types = NULL
)
}
\arguments{
\item{xlfile}{The full file path to an Excel file structured for SLAM data}

\item{BinWidth}{The width of the bins for the catch-at-weight data. Only needed if the CAW data is raw.}

\item{use_Frwpen}{Use the penalty for the random walk in fishing mortality?}

\item{use_R0rwpen}{Use the penalty for the random walk in monthly seasonal recruitment pattern?}

\item{...}{Additional parameters (not currently used)}

\item{Sampled_Data}{An object of class \code{Simulated} generated by \code{Generate_Data}}

\item{sim}{Simulation number}

\item{Fit_Effort}{Logical. Fit effort data?}

\item{Fit_CPUE}{Logical. Fit CPUE data?}

\item{Fit_CAW}{Logical. Fit catch-at-weight data?}

\item{Fit_CAA}{Logical. Fit catch-at-age data?}

\item{Data_types}{Optional character vector specifying data types to fit to.}
}
\value{
A \code{Data} object for the \code{Assess} function
}
\description{
Import and organize data for assessment model
}
\section{Methods (by class)}{
\itemize{
\item \code{Import_Data(Simulated)}: Import simulated data

}}
