% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import_Parameters.R
\name{Import_LifeHistory}
\alias{Import_LifeHistory}
\alias{Import_Exploitation}
\alias{Import_Data}
\title{Import the Life History Parameters from a CSV file}
\usage{
Import_LifeHistory(csv = "Parameters.csv", dir = getwd())

Import_Exploitation(csv = "Parameters.csv", dir = getwd())

Import_Data(csv = "Parameters.csv", dir = getwd())
}
\arguments{
\item{csv}{The name of the CSV file containing the parameters}

\item{dir}{The directory the CSV is in (defaults to current working directory)}
}
\description{
Import the Life History Parameters from a CSV file

Import the Exploitation Parameters from a CSV file

Import the Data Parameters from a CSV file
}
