% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Assess.R
\name{Initialize_Parameters}
\alias{Initialize_Parameters}
\title{Iniitialize Parameters for SLAM Estimation}
\usage{
Initialize_Parameters(
  SimMod,
  log_sigmaF = log(0.05),
  log_sigmaR0 = log(0.1),
  log_sigmaR = log(0.3)
)

Initialize_Parameters(
  SimMod,
  log_sigmaF = log(0.05),
  log_sigmaR0 = log(0.1),
  log_sigmaR = log(0.3)
)
}
\arguments{
\item{log_sigmaR}{}

\item{data}{A data object}

\item{as50}{initial value for age at 50\% selection}

\item{as95}{initial value for age at 95\% selection}
}
\description{
Iniitialize Parameters for SLAM Estimation

Title
}
