% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Assess.R
\name{Initialize_Parameters_OM}
\alias{Initialize_Parameters_OM}
\title{Title}
\usage{
Initialize_Parameters_OM(
  SimMod,
  log_sigmaF = log(0.05),
  log_sigmaR0 = log(0.1),
  log_sigmaR = log(0.3)
)
}
\arguments{
\item{log_sigmaR}{}
}
\description{
Title
}
