% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import_Parameters.R
\docType{data}
\name{LifeHistory}
\alias{LifeHistory}
\title{Example Life History Parameters}
\format{
\subsection{\verb{Life History}}{

A named list:
\describe{
\item{Stock_Name}{A name for the }
\item{Species}{}
\item{maxage}{}
}
}
}
\usage{
LifeHistory
}
\description{
Example Life History Parameters
}
\keyword{datasets}
