% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Process_case_study.R
\name{Make_Summary_Table}
\alias{Make_Summary_Table}
\title{Summarize Case Study Data}
\usage{
Make_Summary_Table(data)
}
\arguments{
\item{data}{a data.frame returned by \code{Process_Data}}
}
\value{
A data.frame
}
\description{
Summarize Case Study Data
}
