% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Report.R
\name{plot_at_Age}
\alias{plot_at_Age}
\alias{plot_weight_at_age}
\alias{plot_maturity_at_age}
\alias{plot_M_at_age}
\alias{plot_PSM_at_age}
\title{Plot the at-age schedules}
\usage{
plot_at_Age(data)

plot_weight_at_age(data)

plot_maturity_at_age(data)

plot_M_at_age(data)

plot_PSM_at_age(data)
}
\arguments{
\item{data}{An object of class \code{Data}}
}
\value{
A \code{ggplot} object
}
\description{
Plot the at-age schedules
}
\section{Functions}{
\itemize{
\item \code{plot_weight_at_age()}: Plot mean weight-at-age with 10th and 90th percentiles

\item \code{plot_maturity_at_age()}: Plot maturity-at-age schedule

\item \code{plot_M_at_age()}: Plot natural mortality-at-age schedule

\item \code{plot_PSM_at_age()}: Plot post-spawning mortality-at-age schedule

}}
\examples{
Data <- Import(Example_Data()[1])
plot_at_Age(Data)
}
