% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Report.R
\name{plot_CAW}
\alias{plot_CAW}
\title{Plot the weight composition data, and model fits to the data}
\usage{
plot_CAW(data)
}
\arguments{
\item{data}{An object of class \code{Data} or class \code{Assess}}
}
\value{
A \code{list} with a \code{data.frame} and  a \code{ggplot} object
}
\description{
Plot the weight composition data, and model fits to the data
}
\examples{
Data <- Import(Example_Data()[1])
plot_CAW(Data)$p
}
