% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Report.R
\name{plot_F}
\alias{plot_F}
\title{Plot the predicted fishing mortality (F)}
\usage{
plot_F(assess)
}
\arguments{
\item{assess}{An object of class \code{Assess}}
}
\value{
A list with a data.frame and a ggplot object
}
\description{
Plot the predicted fishing mortality (F)
}
\examples{
MyData <- Import(Example_Data()[2])
MyAssess <- Assess(MyData)
plot_F(MyAssess)
}
