% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Report.R
\name{plot_Indices}
\alias{plot_Indices}
\alias{plot_Biomass}
\alias{plot_Effort}
\alias{plot_Catch}
\title{Plot indices of relative biomass, effort, and catch}
\usage{
plot_Indices(data)

plot_Biomass(data)

plot_Effort(data)

plot_Catch(data)
}
\arguments{
\item{data}{An object of class \code{Data} or class \code{Assess}}
}
\value{
A \verb{ggplot object}
}
\description{
Plot indices of relative biomass, effort, and catch
}
\section{Functions}{
\itemize{
\item \code{plot_Biomass()}: Plot Index of Biomass time-series

\item \code{plot_Effort()}: Plot Index of Effort time-series

\item \code{plot_Catch()}: Plot Catch of Effort time-series

}}
\examples{
Data <- Import(Example_Data()[2])
plot_Indices(Data)
}
