% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Report.R
\name{plot_seasonal_rec}
\alias{plot_seasonal_rec}
\alias{get_seasonal_recruitment}
\title{Plot the predicted seasonal recruitment pattern}
\usage{
plot_seasonal_rec(assess)

get_seasonal_recruitment(assess)
}
\arguments{
\item{assess}{An object of class \code{Assess}}
}
\value{
A list with a data.frame and a ggplot object
}
\description{
Plot the predicted seasonal recruitment pattern
}
\section{Functions}{
\itemize{
\item \code{get_seasonal_recruitment()}: get data.frame of predicted seasonal recruitment

}}
\examples{
MyData <- Import(Example_Data()[2])
MyAssess <- Assess(MyData)
plot_seasonal_rec(MyAssess)
}
