% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Report.R
\name{plot_SPR}
\alias{plot_SPR}
\title{Plot the predicted spawning potential ratio (SPR)}
\usage{
plot_SPR(assess)
}
\arguments{
\item{assess}{An object of class \code{Assess}}
}
\value{
A list with a data.frame and a ggplot object
}
\description{
Plot the predicted spawning potential ratio (SPR)
}
\examples{
MyData <- Import(Example_Data()[2])
MyAssess <- Assess(MyData)
plot_SPR(MyAssess)
}
