% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Process_case_study.R
\name{Process_Data}
\alias{Process_Data}
\title{Process Case Study Data}
\usage{
Process_Data(village, Case_Study_Data)
}
\arguments{
\item{village}{Name of the village}

\item{Case_Study_Data}{A data.frame}
}
\value{
A list
}
\description{
Process Case Study Data
}
