% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Report.R
\name{Report}
\alias{Report}
\alias{Report.Data}
\alias{Report.Assess}
\title{Generate Data and Assessment Reports}
\usage{
Report(
  x,
  filename = "Report",
  dir = tempdir(),
  open_file = TRUE,
  silent = FALSE,
  ...
)

\method{Report}{Data}(
  x,
  filename = "Data_Report",
  dir = tempdir(),
  open_file = TRUE,
  silent = FALSE,
  ...
)

\method{Report}{Assess}(
  x,
  filename = "Assessment_Report",
  dir = tempdir(),
  open_file = TRUE,
  silent = FALSE,
  ...
)
}
\arguments{
\item{x}{An object of class \code{Data} or \code{Assess}}

\item{filename}{The name of the output file. Defaults to 'Report.html'}

\item{dir}{The directory to save the report. Defaults to a temporary directory.}

\item{open_file}{Logical. Open the file in the browser?}

\item{silent}{Logical. Hide messages?}

\item{...}{Other arguments (currently not used)}
}
\description{
Generates a HTML report and opens it in the web browser
}
