% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Report.R
\name{report_NLL}
\alias{report_NLL}
\title{Report the negative log-likelihoods}
\usage{
report_NLL(assess)
}
\arguments{
\item{assess}{An object of class \code{Assess}}
}
\value{
A data.frame
}
\description{
Report the negative log-likelihoods
}
\examples{
MyData <- Import(Example_Data()[2])
MyAssess <- Assess(MyData)
report_NLL(MyAssess)
}
