% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim_Testing.R
\name{Sim_Test}
\alias{Sim_Test}
\title{Run a Simulation Test}
\usage{
Sim_Test(x, grid, Simulation, Sampling)
}
\arguments{
\item{x}{Scenario number. Row of \code{grid}}

\item{grid}{data.frame with simulation grid. Columns \code{n_months} and \code{data_types}}

\item{Simulation}{An object of class \code{Simulation}}

\item{Sampling}{An object of class \code{Sampling}}
}
\value{
A list
}
\description{
Run a Simulation Test
}
