% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopDynamics.R
\name{Simulate}
\alias{Simulate}
\title{Simulate the fishery}
\usage{
Simulate(Pars, sim = 1)
}
\arguments{
\item{Pars}{A list of parameters}
}
\value{
A dataframe with simulated fishery dynamics
}
\description{
Simulate the fishery
}
