% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SLAM-package.R
\docType{package}
\name{SLAM-package}
\alias{SLAM}
\alias{SLAM-package}
\title{SLAM: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
