% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Process_case_study.R
\name{Summarize_Data}
\alias{Summarize_Data}
\title{Summarize Case Study Data}
\usage{
Summarize_Data(village, Case_Study_Data, Case_Study_Sites)
}
\arguments{
\item{data}{a data.frame returned by \code{Process_Data}}
}
\value{
A data.frame
}
\description{
Summarize Case Study Data
}
