% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_gethr.R
\name{surv_gethr}
\alias{surv_gethr}
\title{Get Hazard Ratio from Cox Model}
\usage{
surv_gethr(fit, vars, names = NULL, ndec = 2)
}
\arguments{
\item{fit}{coxph model}

\item{vars}{the covariates fo the cox model that we would like to extract hazard ratios for}

\item{names}{character vector of same length as vars that provides formal names for the variables
in place of variable name.}

\item{ndec}{numeric that specifies the number of decimal places to round the hazard ratios}
}
\value{
\item{hr_table}{
 a dataframe whos rows correspond the variables inputted by the vars argument and whos names are
 determined by the names argument. If the names argument is not given the names are the same
 as the variable names.
 This dataframe has the following columns (1) value - provides the hazard ratio
 (2) lower - lower bound of 95% confidence interval for hazard ratio
 (3) upper - upper bound of 95% confidence interval for hazard ratio
 (4) pval - the p-value for the hazard ratio
 (5) final - a character stirng that includes columns 1-4 for table output
 (6) with_names - character string like final, but the variable name is included
 }
 \item{hr_text}{
 a character string that has the with names for each variable separated by \\n,
 so that the hazard ratios for vars can be included in a kaplan meier plot.
 }
}
\description{
The purpose of this function is so that the user can easily extract hazard ratios,
hazard ratio confidence intervals, and hazard ratio pvalues easily. This function
also outputs Hazard Ratios that can easily be added plots
}
