% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_mfreq.R
\name{vis_mfreq}
\alias{vis_mfreq}
\title{Visualize Measurement Frequency}
\usage{
vis_mfreq(data, id, ...)
}
\arguments{
\item{data}{a dataframe in long format}

\item{id}{a character string specifying subject id}

\item{...}{optional parameters to pass to barplot function}
}
\value{
prints a barplot
}
\description{
Creates a barplot of measurement frequencies by some id column
}
\seealso{
\link[graphics]{barplot}
}
\author{
William Mueller
}
