% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulate.R
\name{Write_Data2XL}
\alias{Write_Data2XL}
\title{Write Simulated Data to an Excel File}
\usage{
Write_Data2XL(
  Simulation,
  xlfile,
  n_years = 3,
  sim = 1,
  Sampling = NULL,
  binned = FALSE,
  dev = F
)
}
\arguments{
\item{Simulation}{An object of class \code{Simulation}}

\item{xlfile}{The full path (including extension) of Excel file to write}

\item{n_years}{The number of years of data to use. Not used if \code{Sampling} is provided}

\item{sim}{Simulation number}

\item{Sampling}{An optional object of class \code{Simulation}. Defaults to \code{SLAM::Simulation}}

\item{binned}{Logical. Should the CAW data be binned?}

\item{dev}{Logical. Development?}
}
\value{
\code{xlfile} is written to disk
}
\description{
Write Simulated Data to an Excel File
}
