% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_delta.R
\name{add_delta}
\alias{add_delta}
\title{Add delta variable}
\usage{
add_delta(
  data,
  cols,
  id,
  time,
  fill = 0,
  n = 1L,
  type = "lag",
  prefix = paste("delta", type, n, sep = "_")
)
}
\arguments{
\item{data}{the dataframe to be called}

\item{cols}{the character string or character vector of columns to apply this
function to}

\item{id}{the name of the id variable}

\item{time}{variable denoting the time variable}

\item{fill}{numeric or NA to}

\item{n}{number of steps to alter roling difference window}

\item{type}{character string whether lead or lag should be used}

\item{prefix}{prefix denoting type of delta}
}
\description{
This function will create a delta variable given a subject id, variable name
and dataframe.
}
