% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_delta.R
\name{add_delta}
\alias{add_delta}
\title{Add delta variable}
\usage{
add_delta(
  data,
  cols,
  id,
  time,
  fill = 0,
  n = 1L,
  type = "lag",
  prefix = paste("delta", type, n, sep = "_")
)
}
\arguments{
\item{data}{the dataframe to be called}

\item{cols}{the character string or character vector of columns to apply this
function to}

\item{id}{a character string specifying the name of the id variable}

\item{time}{a character stirng the time variable}

\item{fill}{numeric or NA to specifying what to fill the values that fall
outside the offset window. For example, if we ar taking lagged differences, this
will set the value of the firt observation.}

\item{n}{number of steps to alter rolling difference window}

\item{type}{character string whether lead or lag should be used}

\item{prefix}{a character string denothing the prefix of the delta column.
The column name for the new delta column will be "prefix_oldname"}
}
\description{
This function will create a delta variable given a subject id, variable name
and dataframe. It will also skip NA values in the calculation of the delta.
}
\examples{

# Example ----------------------------------------------------------------------

# dplyr must be in NAMESPACE
if (requireNamespace("dplyr", quietly = TRUE)) {

  # merge nmr with SLAM census
  data <- dplyr::left_join(data_SLAM_nmr, data_SLAM_census, by = "idno")

  # checkout data
  head(data)

  # add delta variables with 1 time lagged differences and fill with 0
  data <- add_delta(data = data,
                    cols = c("bw", "fat", "lean", "fluid"),
                    id = "idno",
                    time = "date")

  # add delta variable with 2 time lagged differences to same dataset and fill with 0
  data <- add_delta(data = data,
                    cols = c("bw", "fat", "lean", "fluid"),
                    id = "idno",
                    time = "date",
                    n = 2)
}


}
\author{
Jorge Martinez Romero, William Mueller
}
