% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Assess.R
\name{Assess}
\alias{Assess}
\title{Title}
\usage{
Assess(
  data,
  as50 = 4,
  as95 = 6,
  Feq_init = 0.05,
  sigmaR = 0.5,
  control = list(eval.max = 20000, iter.max = 20000, abs.tol = 1e-20),
  map = list(log_sigmaF = factor(NA), log_sigmaR = factor(NA), log_sigmaR0 = factor(NA)),
  log_sigmaF = log(0.05),
  log_sigmaR0 = log(0.1)
)
}
\arguments{
\item{log_sigmaR0}{}
}
\description{
Title
}
