% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Assess.R
\name{Assess}
\alias{Assess}
\title{Title}
\usage{
Assess(
  Data,
  Parameters = NULL,
  control = list(eval.max = 20000, iter.max = 20000, abs.tol = 1e-20),
  ...
)
}
\arguments{
\item{Data}{A \code{Data} object}

\item{Parameters}{A \code{Parameters} object}

\item{control}{Optional controls for optimizer}

\item{...}{Additional arguments pass to TMB}
}
\description{
Title
}
