% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Assess.R
\name{Assess}
\alias{Assess}
\title{Conduct an Assessment}
\usage{
Assess(
  Data,
  Parameters = NULL,
  max_ESS = 200,
  Est_Rec_Devs = ifelse(length(Data$Year) >= 24, TRUE, FALSE),
  Est_Seasonal = TRUE,
  control = list(eval.max = 20000, iter.max = 20000, abs.tol = 1e-20),
  ...
)
}
\arguments{
\item{Data}{A \code{Data} object}

\item{Parameters}{A \code{Parameters} object}

\item{max_ESS}{Maximum effective sample size for the catch-at-weight data}

\item{Est_Rec_Devs}{Logical. Estimate recruitment deviations (TRUE) or assume constant recruitment (FALSE)}

\item{control}{Optional controls for optimizer}

\item{...}{Additional arguments pass to TMB}
}
\value{
A list of class \code{Assess}
}
\description{
Conduct an Assessment
}
