% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimize_F_Pattern.R
\name{calculate_optimal_fishing}
\alias{calculate_optimal_fishing}
\title{Title}
\usage{
calculate_optimal_fishing(
  LifeHistory,
  Exploitation,
  opt_type = 1,
  utilpow = 0.3
)
}
\arguments{
\item{utilpow}{}
}
\description{
Title
}
