% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Assess.R
\name{Construct_Data_OM}
\alias{Construct_Data_OM}
\title{Title}
\usage{
Construct_Data_OM(
  sim = 1,
  SimMod,
  CAW_Monthly_ESS = 100,
  CAA_Monthly_ESS = 100,
  Effort_CV = 0.2,
  CPUE_CV = 0.2,
  Fit_Effort = 1,
  Fit_CPUE = 1,
  Fit_CAW = 1,
  Fit_CAA = 1,
  use_Frwpen = 1,
  use_R0rwpen = 1,
  Data_types = NULL
)
}
\arguments{
\item{use_Fmeanprior}{}
}
\description{
Title
}
