% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_SLAM_census.R
\docType{data}
\name{data_SLAM_census}
\alias{data_SLAM_census}
\title{SLAM Census}
\format{
a dataframe 2783 obs and 12 variables
\describe{
  \item{cohort}{cohort of the mouse}
  \item{animal_id}{unique identification number for vivarium}
  \item{tag}{unique identifier for each mouse}
  \item{taghistory}{history of unique identifiers for each mouse, they sometimes have changed
  over the course of the study}
  \item{sex}{M or F denoting Male or Female}
  \item{dob}{date variable denoting date of birth}
  \item{idno}{another unique identifier from 1 to n, where 1 is the first mouse enrolled and n
  is the most recent mouse enrolled}
  \item{cage}{the number denoting the cage. The mice were housed mostly in groups of 4}
  \item{eartag}{marking pattern on mouses ears so animal technicians could identify the mouse}
  \item{name}{random name given to each mouse}
}
}
\usage{
data(data_SLAM_census)
}
\description{
A dataset that contains census information for each mouse in SLAM
}
\references{
Palliyaguru DL, Shiroma EJ, Nam JK, Duregon E, Vieira Ligo Teixeira C, 
Price NL, Bernier M, Camandola S, Vaughan KL, Colman RJ, Deighan A, 
Korstanje R, Peters LL, Dickinson SL, Ejima K, Simonsick EM, Launer LJ, 
Chia CW, Egan J, Allison DB, Churchill GA, Anderson RM, Ferrucci L, 
Mattison JA, de Cabo R. Fasting blood glucose as a predictor of 
mortality: Lost in translation. Cell Metab. 2021 Nov 2;33(11):2189-2200.e3. 
doi: 10.1016/j.cmet.2021.0#` 8.013. Epub 2021 Sep 10. PMID: 34508697.

Palliyaguru DL, Vieira Ligo Teixeira C, Duregon E, di Germanio C, Alfaras 
I, Mitchell SJ, Navas-Enamorado I, Shiroma EJ, Studenski S, Bernier M, 
Camandola S, Price NL, Ferrucci L, de Cabo R. Study of Longitudinal Aging 
in Mice: Presentation of Experimental Techniques. J Gerontol A Biol Sci Med 
Sci. 2021 Mar 31;76(4):552-560. doi: 10.1093/gerona/glaa285. 
PMID: 33211821; PMCID: PMC8011709.
}
\keyword{datasets}
