% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_SLAM_nmr.R
\docType{data}
\name{data_SLAM_nmr}
\alias{data_SLAM_nmr}
\title{SLAM NMR Data}
\format{
a dataframe 17467 obs and 6 variables
\describe{
  \item{idno}{unique identifier}
  \item{date}{date of nmr}
  \item{bw}{Body weight in grams}
  \item{fat}{fat mass in grams}
  \item{lean}{lean mass in grams}
  \item{fluid}{fluid mass in grams}
}
}
\usage{
data(data_SLAM_nmr)
}
\description{
A dataset that includes Body Weight, Body Fat, Lean Mass, Fluid Mass for SLAM
mice
}
\keyword{datasets}
