% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generate_Monthly_Recruitment.R
\name{Generate_Monthly_Recruitment}
\alias{Generate_Monthly_Recruitment}
\title{Generate Seasonal Recruitment Pattern}
\usage{
Generate_Monthly_Recruitment(mu = 6, sigma = 2)
}
\arguments{
\item{mu}{Mean calendar month for peak recruitment}

\item{sigma}{Standard deviation}
}
\value{
A numeric vector of seasonal recruitment
}
\description{
Generate Seasonal Recruitment Pattern
}
\examples{
# Pulse recruitment in the beginning of the year
plot(1:12, Generate_Monthly_Recruitment(), type='l')

}
