% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import_Data.R
\name{Import}
\alias{Import}
\title{Import and organize data for assessment model}
\usage{
Import(csvfile, silent = FALSE, ...)
}
\arguments{
\item{csvfile}{The full file path to an CSV file structured for SLAM data}

\item{silent}{Logical. Hide the messages?}

\item{...}{Additional parameters (not currently used)}
}
\value{
A \code{Data} object for the \code{Assess} function
}
\description{
Import and organize data for assessment model
}
