% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import_Data.R
\name{Import.Simulated}
\alias{Import.Simulated}
\title{Import Simulated Data}
\usage{
\method{Import}{Simulated}(
  Sampled_Data = NULL,
  sim = 1,
  Fit_Effort = 1,
  Fit_Catch = 1,
  Fit_CPUE = 1,
  Fit_CAW = 1,
  Fit_CAA = 1,
  use_Frwpen = 1,
  use_R0rwpen = 1,
  CAW_ESS = 200,
  Effort_CV = 0.2,
  Catch_CV = 0.2,
  Index_CV = 0.3,
  Data_types = NULL
)
}
\arguments{
\item{Sampled_Data}{An object of class \code{Simulated} generated by \code{Generate_Data}}

\item{sim}{Simulation number}

\item{Fit_Effort}{Logical. Fit effort data?}

\item{Fit_CPUE}{Logical. Fit CPUE data?}

\item{Fit_CAW}{Logical. Fit catch-at-weight data?}

\item{Fit_CAA}{Logical. Fit catch-at-age data?}

\item{use_Frwpen}{Logical. Use the random walk penalty for F?}

\item{use_R0rwpen}{Logical. Use the random walk penalty for seasonal recruitment?}

\item{Data_types}{Optional character vector specifying data types to fit to.}
}
\value{
A \code{Data} object
}
\description{
Import Simulated Data
}
