% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import_Parameters.R
\name{Import_Parameters}
\alias{Import_Parameters}
\alias{Import_LifeHistory}
\alias{Import_Exploitation}
\alias{Import_Sampling}
\title{Import the Life History, Exploitation, and Sampling Parameters}
\usage{
Import_Parameters(csv = NULL, dev = FALSE)

Import_LifeHistory(csv = NULL, dev = FALSE)

Import_Exploitation(csv = NULL, dev = FALSE)

Import_Sampling(csv = NULL, dev = FALSE)
}
\arguments{
\item{csv}{The path and name of the CSV file containing the parameters}
}
\value{
A named list with the life-history, exploitation, and sampling parameters

A named list with the life history parameters

A named list with the exploitation parameters
}
\description{
The Life History, Exploitation, and Sampling Parameters can imported from a correctly
formatted CSV file. See \code{Simulation_Parameters.csv} included with the \code{SLAM} package
(found at \code{File_Location()}) for an example.
}
