% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Assess.R
\name{Initialize_Parameters}
\alias{Initialize_Parameters}
\title{Iniitialize Parameters for SLAM Estimation}
\usage{
Initialize_Parameters(
  data,
  as50 = 4,
  as95 = 6,
  Feq_init = 0.15,
  sigmaR = 0.5,
  sigmaF_m = 0.4,
  sigmaR0 = 0.7
)
}
\arguments{
\item{data}{A data object}

\item{as50}{initial value for age at 50\% selection}

\item{as95}{initial value for age at 95\% selection}

\item{log_sigmaR0}{}
}
\description{
Iniitialize Parameters for SLAM Estimation
}
