% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Assess.R
\name{Initialize_Parameters}
\alias{Initialize_Parameters}
\title{Initialize Parameters for SLAM Estimation}
\usage{
Initialize_Parameters(
  data,
  as50 = 4,
  as95 = 6,
  Feq_init = 0.15,
  F_ts = 0.1,
  sigmaR = 0.6,
  sigmaF_m = 0.4,
  sigmaR0 = 0.4
)
}
\arguments{
\item{data}{An object of class \code{Data}}

\item{as50}{initial value for age at 50\% selection}

\item{as95}{initial value for age at 95\% selection}

\item{Feq_init}{initial value for initial equilibrium fishing mortality}

\item{F_ts}{initial values for monthly fishing mortality. Must be length \code{length(Data$Years)}}

\item{sigmaR}{standard deviation of log-normal recruitment deviations}

\item{sigmaF_m}{standard deviation of random-walk penalty for monthly fishing mortality}

\item{sigmaR0}{standard deviation of random-walk penalty for monthly (seasonal) recruitment}
}
\value{
A list of class \code{Parameters}
}
\description{
Initialize Parameters for SLAM Estimation
}
