% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Assess.R
\name{Initialize_Parameters_OM}
\alias{Initialize_Parameters_OM}
\title{Title}
\usage{
Initialize_Parameters_OM(SimMod, sigmaR = 0.5, sigmaF_m = 0.4, sigmaR0 = 0.5)
}
\arguments{
\item{log_sigmaR}{}
}
\description{
Title
}
