% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Process_case_study.R
\name{Process_Data}
\alias{Process_Data}
\title{Process Case Study Data}
\usage{
Process_Data(Sites, Case_Study_Data, maxWeightBin = 4, weight_bin = 0.25)
}
\arguments{
\item{Sites}{Names of the village}

\item{Case_Study_Data}{A data.frame}
}
\value{
A list
}
\description{
Process Case Study Data
}
