% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulate.R
\name{Simulate}
\alias{Simulate}
\title{Simulate a fishery for a short-lived semelparous species}
\usage{
Simulate(
  LifeHistory,
  Exploitation,
  Data,
  nsim = 3,
  seed = 101,
  currentYr = format(Sys.Date(), "\%Y")
)
}
\arguments{
\item{LifeHistory}{A \code{list} object with the life-history parameters}

\item{Data}{A \code{list} object with the parameters for generating data}

\item{nsim}{The number of stochastic simulations}

\item{seed}{The seed for the random number generator}

\item{currentYr}{The current (most recent) year for the simulations}

\item{Explotation}{A \code{list} object with the exploitation parameters}
}
\description{
Simulate a fishery for a short-lived semelparous species
}
