% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SLAM-package.R
\docType{package}
\name{SLAM-package}
\alias{SLAM}
\alias{SLAM-package}
\title{SLAM: SLAM: Short-Lived Assessment Model}
\description{
A model to simulate and assess short-lived semelparous species such as octopus.
}
\author{
\strong{Maintainer}: Adrian Hordyk \email{adrian@bluematterscience.com} (\href{https://orcid.org/0000-0001-5620-3446}{ORCID})

}
\keyword{internal}
